refactor(ai): extract shared chat-client tracing into Granit.AI
Hoist the per-provider tracing decorator logic into a shared TracingChatClient
and AIChatTracingProfile in Granit.AI; the Anthropic, AzureOpenAI, Ollama, and
OpenAI Tracing*ChatClient decorators now delegate to it (~590 lines removed).
Each provider ActivitySource gains the profile wiring.
