+++ src/Granit.AI.OpenAI/Diagnostics/AIOpenAIActivitySource.cs
@@ -1,4 +1,5 @@
 using System.Diagnostics;
+using Granit.AI.Diagnostics;
 
 namespace Granit.AI.OpenAI.Diagnostics;
 
@@ -26,4 +27,8 @@ internal static class AIOpenAIActivitySource
 
     /// <summary>The singleton <see cref="ActivitySource"/> instance.</summary>
     internal static readonly ActivitySource Source = new(Name);
+
+    /// <summary>Tracing identity supplied to <see cref="TracingChatClient"/>.</summary>
+    internal static readonly AIChatTracingProfile TracingProfile =
+        new(Source, SystemTagValue, ChatOperation, ChatStreamOperation);
 }
